Restrict nginx upstream to containers on the same node.
When running one service on multiple clusters in different regions, the nginx node would pick up all containers of the service as upstream. This might lead to inter-region communication. This change restricts the upstream containers to containers on the same node as the nginx proxy itself. Running nginx on all nodes using this method might give a better bandwidth even though the load balancing has to be done by a 3rdparty service instead of tutum.