Added restriction mode.
The restriction mode can be one of the following values: node, region or none.

node: restricts upstream to containers on the same node as nginx-proxy.
region: restricts upstream to containers on the same region. This maintains the advantage of private connections between nodes in the same datacenter. Currently multiple custom nodes are treated as a single nil region.
none: guess it :)